# [SPARK-49178][SQL] Optimize performance of `Row#getSeq` to match the performance when using Spark 3.5 with Scala 2.12